Finished rough setup of W-tree construction. NEEDS TESTING

+++ src/main/java/edu/usc/WinnTree/Construction/AffinityCalculation/VisualScore.java
@@ -7,8 +7,8 @@ import java.util.List;
 public class VisualScore {
 
     public static double CalculateVisualScore(FunctionalArea vertexOne, FunctionalArea vertexTwo){
-        List<Float> vertexOne_centroid = vertexOne.getCentroid();
-        List<Float> vertexTwo_centroid = vertexTwo.getCentroid();
+        List<Double> vertexOne_centroid = vertexOne.getCentroid();
+        List<Double> vertexTwo_centroid = vertexTwo.getCentroid();
         double hypontenuse = Math.hypot(vertexOne_centroid.get(0) - vertexTwo_centroid.get(0), vertexOne_centroid.get(1) - vertexTwo_centroid.get(1));
 
         //Need to scale visual distance to {0 .... 1)
